bg-removal-backend/requirements.txt
flask==3.0.0
rembg[new]==2.0.50
pillow==10.1.0
numpy==1.24.3
torch==2.1.0
torchvision==0.16.0
gunicorn==21.2.0
onnxruntime-gpu==1.16.3

